pagecache
Struct Stack

pub struct Stack<T: Send + 'static> { /* private fields */ }

A simple lock-free stack, with the ability to atomically append or entirely swap-out entries.

Implementations

impl<T: Clone + Send + Sync + 'static> Stack<T>

pub fn push(&self, inner: T)

Add an item to the stack, spinning until successful.

pub fn cap<'g>( &self, old: Shared<'_, Node<T>>, new: T, guard: &'g Guard, ) -> Result<Shared<'g, Node<T>>, (Shared<'g, Node<T>>, Owned<Node<T>>)>

compare and push

pub fn cap_node<'g>( &self, old: Shared<'_, Node<T>>, node: Owned<Node<T>>, guard: &'g Guard, ) -> Result<Shared<'g, Node<T>>, (Shared<'g, Node<T>>, Owned<Node<T>>)>

compare and push

pub fn cas<'g>( &self, old: Shared<'g, Node<T>>, new: Owned<Node<T>>, guard: &'g Guard, ) -> Result<Shared<'g, Node<T>>, (Shared<'g, Node<T>>, Owned<Node<T>>)>

compare and swap

pub fn head<'g>(&self, guard: &'g Guard) -> Shared<'g, Node<T>>

Returns the current head pointer of the stack, which can later be used as the key for cas and cap operations.